# savicar-api/docs/swagger.yaml
basePath: /
definitions:
  badencar-api_internal_domain_services.Service:
    properties:
      code:
        type: string
      description:
        type: string
      hours_quantity:
        type: number
      id_order:
        type: integer
      id_service:
        type: integer
      total_value:
        type: number
      unit_value:
        type: number
    type: object
...
  serviceorder.ServiceOrder:
    properties:
      customer_name:
        type: string
      customer_notes:
        type: string
      date_time_in:
        type: string
      date_time_out:
        type: string
      diagnosis_notes:
        type: string
      discount:
        type: number
      final_amount:
        type: number
      id_customer:
        type: integer
      id_customer_model:
        type: integer
      id_order:
        type: integer
      id_technician:
        type: integer
      internal_notes:
        type: string
      model_name:
        type: string
      odometer_reading:
        type: integer
      plate_number:
        type: string
      service_type:
        type: integer
      technician_name:
        type: string
      total_amount:
        type: number
      vin:
        type: string
    type: object
  serviceorderimage.ServiceOrderImage:
    properties:
      id_image:
        type: integer
      id_order:
        type: integer
      image_path:
        type: string
      video_path:
        type: string
    type: object
...
  technician.Technician:
    properties:
      id_technician:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Service Order Images API
  title: Badencar API
  version: "1.0"
paths:
...
  /service-order-images/order/{id_order}:
    get:
      parameters:
      - description: Service Order ID
        in: path
        name: id_order
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/serviceorderimage.ServiceOrderImage'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List images metadata for an order
      tags:
      - service-order-images
...
  /service-orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/serviceorder.ServiceOrder'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all service orders
      tags:
      - service-orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/serviceorder.ServiceOrder'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/serviceorder.ServiceOrder'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a service order
      tags:
      - service-orders
  /service-orders/{id}:
    delete:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a service order
      tags:
      - service-orders
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serviceorder.ServiceOrder'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a service order by ID
      tags:
      - service-orders
    put:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/serviceorder.ServiceOrder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serviceorder.ServiceOrder'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a service order
      tags:
      - service-orders
  /services:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/badencar-api_internal_domain_services.Service'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all services
      tags:
      - services
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/badencar-api_internal_domain_services.Service'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/badencar-api_internal_domain_services.Service'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a service
      tags:
      - services
  /services/{id}:
    delete:
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a service
      tags:
      - services
    get:
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/badencar-api_internal_domain_services.Service'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a service by ID
      tags:
      - services
    put:
      consumes:
      - application/json
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/badencar-api_internal_domain_services.Service'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/badencar-api_internal_domain_services.Service'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a service
      tags:
      - services
  /technicians:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/technician.Technician'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all technicians
      tags:
      - technicians
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/technician.Technician'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/technician.Technician'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a technician
      tags:
      - technicians
  /technicians/{id}:
    delete:
      parameters:
      - description: Technician ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a technician
      tags:
      - technicians
    get:
      parameters:
      - description: Technician ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/technician.Technician'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a technician by ID
      tags:
      - technicians
    put:
      consumes:
      - application/json
      parameters:
      - description: Technician ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/technician.Technician'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/technician.Technician'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a technician
      tags:
      - technicians
...
swagger: "2.0"